signals
+ added signals
    - accessed by holding the grave key and letting go when wanted (you can use your mouse for selection also)
    - you can add your own signals by placing them in `assets/<namespace>/animations/splatcraft/signals/<path>.json`
        - use the `player_classic`/`player_slim` geckolib models used by the mod to create animations
        - signal icons for the selection screen are stored in `assets/<namespace>/textures/splatcraft/gui/signals/<path>.png`
     - you can whitelist signals for moderation if a whitelist is present at `data/<namespace>/splatcraft/signal_whitelist.json`
         - this is an array of strings of signal IDs (`[ "splatcraft:booyah", "splatcraft:come_on" ]`)
* slightly optimised various ink color related functionality
* tweaked movement speed & jump height
* fixed being able to specifically sprint into squid form to gain immense speed
* updated dependencies

+++ src/main/java/com/cibernet/splatcraft/mixin/client/GameRendererMixin.java
@@ -39,7 +39,7 @@ public class GameRendererMixin {
     private float modifyFovForSquidForm(float c) {
         ClientPlayerEntity player = this.client.player;
         if (player != null && SplatcraftConfig.UI.modifyFovForSquidForm.value && PlayerDataComponent.isSquid(player)) {
-            return c + (float) SplatcraftConfig.UI.fovForSquidForm.value / 100;
+            return c + (SplatcraftConfig.UI.fovForSquidForm.value / 100f);
         }
 
         return c;
